ueba_eda/pyproject.toml
[project]
name = "ueba_eda"
version = "0.1.1"
description = "UEBA EDA and Preprocessing Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mypy>=1.19.1",
    "polars>=1.38.1",
    "pyarrow>=23.0.0",
    "pyyaml>=6.0.3",
]

[tool.mypy]
disable_error_code = ["import-untyped"]
